+++ packages/store/tests/plugins.spec.ts
@@ -1,32 +1,38 @@
+import { assertInInjectionContext } from '@angular/core';
 import { TestBed } from '@angular/core/testing';
-import { NgxsModule, NGXS_PLUGINS, Store } from '@ngxs/store';
-import { tap } from 'rxjs/operators';
-import { Observable } from 'rxjs';
+import { NgxsModule, NGXS_PLUGINS, Store, NgxsNextPluginFn, InitState } from '@ngxs/store';
+import { debounceTime, firstValueFrom, tap } from 'rxjs';
 
 describe('Plugins', () => {
-  it('should run a function plugin', () => {
-    let pluginInvoked = 0;
+  it('should run a function plugin (within an injection context too)', async () => {
+    // Arrange
+    const recorder: any[] = [];
 
     class Foo {
       static readonly type = 'Foo';
     }
 
-    function logPlugin(
-      state: any,
-      action: any,
-      next: (state: any, action: any) => Observable<any>
-    ) {
-      if (action.constructor && action.constructor.type === 'Foo') {
-        pluginInvoked++;
+    function asyncLogPlugin(state: any, action: any, next: NgxsNextPluginFn) {
+      assertInInjectionContext(asyncLogPlugin);
+
+      if (action.constructor.type === 'Foo') {
+        recorder.push(['asyncLogPlugin()', action, 'before next()']);
       }
 
       return next(state, action).pipe(
+        debounceTime(0),
         tap(() => {
           if (action.constructor.type === 'Foo') {
-            pluginInvoked++;
+            recorder.push(['asyncLogPlugin()', action, 'after next()']);
           }
         })
       );
+    }
+
+    function otherPlugin(state: any, action: any, next: NgxsNextPluginFn) {
+      assertInInjectionContext(otherPlugin);
+      recorder.push(['otherPlugin()', action]);
+      return next(state, action);
     }
 
     TestBed.configureTestingModule({
@@ -34,15 +40,33 @@ describe('Plugins', () => {
       providers: [
         {
           provide: NGXS_PLUGINS,
-          useValue: logPlugin,
+          useValue: asyncLogPlugin,
+          multi: true
+        },
+        {
+          provide: NGXS_PLUGINS,
+          useValue: otherPlugin,
           multi: true
         }
       ]
     });
 
-    const store: Store = TestBed.inject(Store);
-    store.dispatch(new Foo());
+    // Act
+    const store = TestBed.inject(Store);
 
-    expect(pluginInvoked).toEqual(2);
+    // Assert
+    expect(recorder).toEqual([['otherPlugin()', new InitState()]]);
+
+    // Act
+    const action = new Foo();
+    await firstValueFrom(store.dispatch(action));
+
+    // Assert
+    expect(recorder).toEqual([
+      ['otherPlugin()', new InitState()],
+      ['asyncLogPlugin()', action, 'before next()'],
+      ['otherPlugin()', action],
+      ['asyncLogPlugin()', action, 'after next()']
+    ]);
   });
 });